[sui-fork] Move GraphQL read contracts into gql
Keep the crate root focused on the service API by colocating the private remote-read traits and query types with their GraphQL implementation.
