+++ crates/sui-fork/src/gql/queries.rs
@@ -38,7 +38,7 @@ pub(crate) mod txn_query {
         Transaction as SuiTransaction, TransactionData, VerifiedTransaction,
     };
 
-    use crate::TransactionInfo;
+    use crate::gql::TransactionInfo;
 
     #[derive(cynic::Scalar, Debug, Clone)]
     #[cynic(graphql_type = "Base64")]
@@ -1345,7 +1345,7 @@ pub(crate) mod object_query {
     use sui_types::object::Object;
 
     use super::*;
-    use crate::{ObjectKey as GqlObjectKey, VersionQuery};
+    use crate::gql::{ObjectKey as GqlObjectKey, VersionQuery};
 
     #[derive(cynic::Scalar, Debug, Clone)]
     #[cynic(graphql_type = "SuiAddress")]
